implement encoding of index
